Refactor Airflow pipeline implementation
