+++ elyra/pipeline/component.py
@@ -227,7 +227,7 @@ class ComponentParser(LoggingConfigurable):  # ABC
     component_platform: RuntimeProcessorType = None
     _file_types: List[str] = None
     _parser_class_map: Dict[str, str] = {
-        "APACHE_AIRFLOW": "elyra.pipeline.airflow.component_parser_airflow:AirflowComponentParser",
+        "APACHE_AIRFLOW": "elyra.pipeline.airflow.airflow_component_parser:AirflowComponentParser",
         "KUBEFLOW_PIPELINES": "elyra.pipeline.kfp.component_parser_kfp:KfpComponentParser",
     }
