+++ elyra/tests/pipeline/airflow/test_processor_airflow.py
@@ -27,7 +27,7 @@ import github
 import pytest
 
 from elyra.metadata.metadata import Metadata
-from elyra.pipeline.airflow.processor_airflow import AirflowPipelineProcessor
+from elyra.pipeline.airflow.airflow_processor import AirflowPipelineProcessor
 from elyra.pipeline.parser import PipelineParser
 from elyra.pipeline.pipeline import GenericOperation
 from elyra.pipeline.pipeline_constants import COS_OBJECT_PREFIX
