+++ pyproject.toml
@@ -137,7 +137,7 @@ metadata-tests = "elyra.tests.metadata.test_utils:MetadataTestSchemasProvider"
 
 [project.entry-points."elyra.pipeline.processors"]
 local = "elyra.pipeline.local.processor_local:LocalPipelineProcessor"
-airflow = "elyra.pipeline.airflow.processor_airflow:AirflowPipelineProcessor"
+airflow = "elyra.pipeline.airflow.airflow_processor:AirflowPipelineProcessor"
 kfp = "elyra.pipeline.kfp.processor_kfp:KfpPipelineProcessor"
 
 [project.entry-points."elyra.component.catalog_types"]
